First version of the custom Minila QMK firmware
Used the `util/new_keyboard.sh` to scaffold the firmware folder
structure.

Used the `util/docker_build.sh` to build the firmware.

Used my `upload_hex_firmware.sh` to upload the firmware. It worked as
the Pro Micros, in the sense that a `/dev/ttyACM0` only appears when we
reset the board (by shorting the RST and GND pins, or pressing the
button).

The layout has a RESET button on the FN4+Backspace (layer FN4) that
works, I've used it to upload some firmware updates.